gpfs: implement Ganesha backend with manila.share.drivers.ganesha
This is an interim commit, which just introduces the new,
manila.share.drivers.ganesha based helper class and
implements helper dispatch. This code is not functional, as
the helper interface used by the original code is not compatible
with the manila.share.drivers.ganesha.NASHelperBase helper
interface, on which the new class is based.

A subsequent commit will address interface incompability.

Change-Id: I4f7ed670eefef22967cd0c2aff711a0f06c51837